Refactor quality actions into shared helpers
